requirements.txt
streamlit
langchain
langchain-google-genai
langchain-community
sentence-transformers
transformers
tokenizers>=0.19.0,<0.22.0
python-dotenv
faiss-cpu
pydantic
plotly
pandas
